requirements.txt
aiofiles==0.4.0
aiohttp==3.6.2
async-timeout==3.0.1
asyncpg==0.19.0
attrs==19.3.0
basictracer==3.1.0
certifi==2019.9.11
chardet==3.0.4
h11==0.8.1
h2==3.1.1
hpack==3.0.0
httpcore==0.3.0
httptools==0.0.13
hyperframe==5.2.0
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
opentracing==2.2.0
peewee==3.11.2
protobuf==3.10.0
psycopg2==2.8.4
python-consul==1.1.0
PyYAML==5.1.2
requests==2.22.0
requests-async==0.5.0
rfc3986==1.3.2
sanic==19.9.0
six==1.13.0
typing-extensions==3.7.4.1
ujson==1.35
urllib3==1.25.7
uvloop==0.14.0
websockets==8.1
yarl==1.3.0
